+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -260,6 +260,18 @@ class SentryOptionsTest {
     }
 
     @Test
+    fun `when options is initialized, memoryCollector is noop`() {
+        assert(SentryOptions().memoryCollector == NoOpMemoryCollector.getInstance())
+    }
+
+    @Test
+    fun `when a null memoryCollector is set, memoryCollector is noop`() {
+        val options = SentryOptions()
+        options.setMemoryCollector(null)
+        assert(SentryOptions().memoryCollector == NoOpMemoryCollector.getInstance())
+    }
+
+    @Test
     fun `when adds scope observer, observer list has it`() {
         val observer = mock<IScopeObserver>()
         val options = SentryOptions().apply {