+++ sentry/src/test/java/io/sentry/SentryTest.kt
@@ -350,8 +350,37 @@ class SentryTest {
         assertTrue { sentryOptions!!.mainThreadChecker is CustomMainThreadChecker }
     }
 
+    @Test
+    fun `overrides memory collector if it's not set`() {
+        var sentryOptions: SentryOptions? = null
+
+        Sentry.init {
+            it.dsn = dsn
+            sentryOptions = it
+        }
+
+        assertTrue { sentryOptions!!.memoryCollector is JavaMemoryCollector }
+    }
+
+    @Test
+    fun `does not override memory collector if it's already set`() {
+        var sentryOptions: SentryOptions? = null
+
+        Sentry.init {
+            it.dsn = dsn
+            it.setMemoryCollector(CustomMemoryCollector())
+            sentryOptions = it
+        }
+
+        assertTrue { sentryOptions!!.memoryCollector is CustomMemoryCollector }
+    }
+
     private class CustomMainThreadChecker : IMainThreadChecker {
         override fun isMainThread(threadId: Long): Boolean = false
+    }
+
+    private class CustomMemoryCollector : IMemoryCollector {
+        override fun collect(): MemoryCollectionData? = null
     }
 
     private class CustomModulesLoader : IModulesLoader {